py-crawler/requirements.txt
pyyaml==6.0.1
requests==2.31.0
schedule==1.2.1
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
deep-translator==1.11.4
vaderSentiment==3.3.2